Unit tests for configToWrapperConfig
