+++ browser/tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
         "outDir": "lib",
-        "rootDir": "./ts",
+        "rootDir": ".",
         "resolveJsonModule": true,
         "module": "commonjs",
         "target": "es5",
